增加产品分类jsp

+++ geekway-service/src/main/java/com/bruce/geekway/service/product/IWxProductSkuRelationService.java
@@ -4,7 +4,7 @@ import com.bruce.foundation.service.IFoundationService;
 import com.bruce.geekway.model.WxProductSkuRelation;
 import com.bruce.geekway.model.WxProductSkuRelationCriteria;
 
-public interface IWxProductSkuRelationService extends IFoundationService<WxProductSkuRelation, Integer, WxProductSkuRelationCriteria>{
+public interface IWxProductSkuRelationService extends IFoundationService<WxProductSkuRelation, Long, WxProductSkuRelationCriteria>{
 
 	/**
 	 * 删除前检查skuPropValueId是否有关联的产品
